backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
python-jose[cryptography]>=3.2.0
passlib[bcrypt]>=1.7.4
bcrypt<4
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.0.0
email-validator>=2.2.0
sqlalchemy>=2.0.0
pandas>=2.1.0
scikit-learn>=1.4.0
joblib>=1.3.0
kagglehub
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.26.0
